added evaluation for deep learning approach
