+++ main.py
@@ -1,4 +1,4 @@
-import scripts.deep.streamlit as st
+import scripts.deep.gpt_resume_optimizer as st
 import pandas as pd
 import re
 import pickle
