+++ requirements.txt
@@ -1,8 +1,131 @@
-streamlit==1.27.2
-pandas==1.5.3
-nltk==3.8.1
-scikit-learn==1.3.0
+et_xmlfile==2.0.0
+executing==2.2.0
+filelock==3.18.0
+fonttools==4.56.0
+fpdf==1.7.2
+frozenlist==1.5.0
+fsspec==2024.12.0
+gitdb==4.0.12
+GitPython==3.1.44
+h11==0.14.0
+httpcore==1.0.7
+httpx==0.28.1
+huggingface-hub==0.29.3
+idna==3.10
+importlib_metadata==8.6.1
+importlib_resources==6.5.2
+ipykernel==6.29.5
+ipython==9.0.2
+ipython_pygments_lexers==1.1.1
+jedi==0.19.2
+jellyfish==1.1.3
+Jinja2==3.1.6
+jiter==0.9.0
+joblib==1.4.2
+jsonschema==4.23.0
+jsonschema-specifications==2024.10.1
+jupyter_client==8.6.3
+jupyter_core==5.7.2
+kagglehub==0.3.10
+kiwisolver==1.4.8
+langcodes==3.5.0
+language_data==1.3.0
+language_tool_python==2.9.0
+lxml==5.3.1
+marisa-trie==1.2.1
+markdown-it-py==3.0.0
+MarkupSafe==3.0.2
+matplotlib==3.10.1
+matplotlib-inline==0.1.7
+mdurl==0.1.2
+mpmath==1.3.0
+multidict==6.2.0
+multiprocess==0.70.16
+murmurhash==1.0.12
+narwhals==1.31.0
+nest-asyncio==1.6.0
+networkx==3.4.2
+numpy==2.2.3
+openai==1.65.3
+openpyxl==3.1.5
+packaging==24.2
+pandas==2.2.3
+parso==0.8.4
+peft==0.15.0
+pexpect==4.9.0
+pillow==11.1.0
+pinecone==6.0.1
+pinecone-plugin-interface==0.0.7
+platformdirs==4.3.6
+preshed==3.0.9
+prompt_toolkit==3.0.50
+propcache==0.3.0
+protobuf==5.29.4
+psutil==7.0.0
+ptyprocess==0.7.0
+pure_eval==0.2.3
+pyarrow==19.0.1
+pyaspeller==2.0.2
+pydantic==2.10.6
+pydantic_core==2.27.2
+pydeck==0.9.1
+Pygments==2.19.1
+PyMuPDF==1.25.3
+pyparsing==3.2.1
+pypdf==5.4.0
 PyPDF2==3.0.1
-python-docx==0.8.11
-kagglehub==1.0.4
-pickle-mixin==1.0.2
+pyphen==0.17.2
+python-dateutil==2.9.0.post0
+python-docx==1.1.2
+python-docx-replace==0.4.4
+python-dotenv==1.0.1
+pytz==2025.1
+PyYAML==6.0.2
+pyzmq==26.3.0
+referencing==0.36.2
+regex==2024.11.6
+requests==2.32.3
+rich==13.9.4
+rpds-py==0.23.1
+safetensors==0.5.3
+scikit-learn==1.6.1
+scipy==1.15.2
+segtok==1.5.11
+sentence-transformers==3.4.1
+setuptools==76.0.0
+shellingham==1.5.4
+six==1.17.0
+smart-open==7.1.0
+smmap==5.0.2
+sniffio==1.3.1
+spacy==3.8.4
+spacy-legacy==3.0.12
+spacy-loggers==1.0.5
+srsly==2.5.1
+stack-data==0.6.3
+streamlit==1.43.2
+sympy==1.13.1
+tabulate==0.9.0
+tenacity==9.0.0
+textstat==0.7.5
+thinc==8.3.4
+threadpoolctl==3.6.0
+tokenizers==0.21.1
+toml==0.10.2
+torch==2.6.0
+tornado==6.4.2
+tqdm==4.67.1
+traitlets==5.14.3
+transformers==4.49.0
+typer==0.15.2
+typing_extensions==4.12.2
+tzdata==2025.1
+urllib3==2.3.0
+wasabi==1.1.3
+wcwidth==0.2.13
+weasel==0.4.1
+wrapt==1.17.2
+xxhash==3.5.0
+yake==0.4.8
+yarl==1.18.3
+zipp==3.21.0
